requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==2.0.2
webdriver-manager==3.8.6
wsproto==1.2.0
